Extract list parsing into a Subparser
The Guides DocumentParser class contains the code for interpreting
everything. This makes for a bit of spaghetti code and makes it harder
to change or understand the impact of changes.

When you observe the DocumentParser, it has a couple of responsibilities
but the most striking are perhaps the ParseLine and Flush method as
these two contain the big switches that do different things based on the
current state of the system.

By creating the concept of a 'subparser' (working title, may change)
that has a parse and build method (matching the parseLine and Flush) I
think it is possible to move the individual handling of state into
subparsers.

Subsequent commits, where the other states will be extracted, will prove
whether this is going to work and will possibly change how a subparser
looks because there is some shared state in the DocumentParser that may
need to be resolved.

The end-goal is to have clear subparser and that we can rework the
subparsers to match the way the RST spec names things and divides
things.
